+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-
+- Upon enabling the plugin, input mappings will be auto generated and the editor will forcibly restart.
+- A new Tree view panel that allows you to set the context by clicking items in the tree. `tree` to open it.
+- Added a `Rename` or `mv` command to rename nodes
 ### Changed
-
+- New UI
+- Fixed input mapping issues
 ### Removed
 
 ## [1.0.0-alpha]